brightmoney_backend/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==8.1.3
Django==3.0.14
django-celery-results==1.2.1
django-environ==0.9.0
django-request-id==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.2
flake8==5.0.4
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
Jinja2==3.1.2
kombu==4.6.10
MarkupSafe==2.1.1
mccabe==0.7.0
plaid-python==4.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2020.1
PyYAML==6.0
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
